harden linux/windows vault implementation for concurrent access. (#3240)
* Harden linux/windows vault implementation for concurrent access

* Spit off some platform-specific code to make linter happy

* Add missing field for windows implementation

* Fix return params for Get on windows

* Replace deprecated ioutil with os package API for windows implementation

* Reorganise platform specific code

* Refactored vault code, extracted shared code, addressed the code review comments

* Split off platform specific options

* Remove wrong build directive from the common options

* do not try to generate HTML test report if test XML output isn't present

---------

(cherry picked from commit c4c314afd643723631e7350e001f308da6438064)

+++ .buildkite/scripts/steps/integration_tests.sh
@@ -13,7 +13,13 @@ TESTS_EXIT_STATUS=$?
 set -e
 
 # HTML report
-go install github.com/alexec/junit2html@latest
-junit2html < build/TEST-go-integration.xml > build/TEST-report.html
+outputXML="build/TEST-go-integration.xml"
+
+if [ -f "$outputXML" ]; then
+  go install github.com/alexec/junit2html@latest
+  junit2html < "$outputXML" > build/TEST-report.html
+else
+    echo "Cannot generate HTML test report: $outputXML not found"
+fi
 
 exit $TESTS_EXIT_STATUS
